refactors config module
