+++ reviews/commands.py
@@ -6,8 +6,9 @@ from rich.console import RenderGroup
 from rich.live import Live
 from rich.panel import Panel
 
-from . import config
-from .config import render_config_table
+from .config import settings
+from .config.controller import render_config_table
+from .config.helpers import get_configuration
 from .controller import GithubPullRequestController, GitlabPullRequestController, PullRequestController
 from .layout import (
     RenderLayoutManager,
@@ -36,10 +37,10 @@ def single_render(provider: str) -> None:
     body = None
 
     if provider == "gitlab":
-        configuration = config.get_configuration(config=config.REVIEWS_GITLAB_REPOSITORY_CONFIGURATION)
+        configuration = get_configuration(config=settings.REVIEWS_GITLAB_REPOSITORY_CONFIGURATION)
         body = GitlabPullRequestController().render(configuration=configuration)
     else:
-        configuration = config.get_configuration(config=config.REVIEWS_GITHUB_REPOSITORY_CONFIGURATION)
+        configuration = get_configuration(config=settings.REVIEWS_GITHUB_REPOSITORY_CONFIGURATION)
         body = GithubPullRequestController().render(configuration=configuration)
 
     layout_manager = RenderLayoutManager(layout=generate_layout(log=False, footer=False))
@@ -75,10 +76,10 @@ def render(provider: str) -> None:
     controller: PullRequestController
 
     if provider == "gitlab":
-        configuration = config.get_configuration(config=config.REVIEWS_GITLAB_REPOSITORY_CONFIGURATION)
+        configuration = get_configuration(config=settings.REVIEWS_GITLAB_REPOSITORY_CONFIGURATION)
         controller = GitlabPullRequestController()
     else:
-        configuration = config.get_configuration(config=config.REVIEWS_GITHUB_REPOSITORY_CONFIGURATION)
+        configuration = get_configuration(config=settings.REVIEWS_GITHUB_REPOSITORY_CONFIGURATION)
         controller = GitlabPullRequestController()
         controller = GithubPullRequestController()
 
@@ -114,7 +115,7 @@ def render(provider: str) -> None:
                 log_component=generate_log_table(logs=logs),
             )
 
-            delay = config.REVIEWS_DELAY_REFRESH * 0.01
+            delay = settings.REVIEWS_DELAY_REFRESH * 0.01
             while not overall_progress.finished:
                 sleep(delay)
                 for job in job_progress.tasks:
@@ -132,33 +133,33 @@ def render_config(show: bool) -> None:
     configurations = [
         {
             "name": "GITHUB_TOKEN",
-            "value": config.GITHUB_TOKEN if show else "".join("*" for _ in range(len(config.GITHUB_TOKEN))),
+            "value": settings.GITHUB_TOKEN if show else "".join("*" for _ in range(len(settings.GITHUB_TOKEN))),
         },
-        {"name": "GITHUB_USER", "value": config.GITHUB_USER},
-        {"name": "GITHUB_URL", "value": config.GITHUB_URL},
+        {"name": "GITHUB_USER", "value": settings.GITHUB_USER},
+        {"name": "GITHUB_URL", "value": settings.GITHUB_URL},
         {
             "name": "GITLAB_TOKEN",
-            "value": config.GITLAB_TOKEN if show else "".join("*" for _ in range(len(config.GITLAB_TOKEN))),
+            "value": settings.GITLAB_TOKEN if show else "".join("*" for _ in range(len(settings.GITLAB_TOKEN))),
         },
-        {"name": "GITLAB_USER", "value": config.GITLAB_USER},
-        {"name": "GITLAB_URL", "value": config.GITLAB_URL},
-        {"name": "REVIEWS_PATH_TO_CONFIG", "value": f"{config.REVIEWS_PATH_TO_CONFIG}"},
+        {"name": "GITLAB_USER", "value": settings.GITLAB_USER},
+        {"name": "GITLAB_URL", "value": settings.GITLAB_URL},
+        {"name": "REVIEWS_PATH_TO_CONFIG", "value": f"{settings.REVIEWS_PATH_TO_CONFIG}"},
         {
             "name": "GITHUB_DEFAULT_PAGE_SIZE",
-            "value": f"{config.GITHUB_DEFAULT_PAGE_SIZE}",
+            "value": f"{settings.GITHUB_DEFAULT_PAGE_SIZE}",
         },
-        {"name": "REVIEWS_DELAY_REFRESH", "value": f"{config.REVIEWS_DELAY_REFRESH}"},
+        {"name": "REVIEWS_DELAY_REFRESH", "value": f"{settings.REVIEWS_DELAY_REFRESH}"},
         {
             "name": "REVIEWS_GITHUB_REPOSITORY_CONFIGURATION",
-            "value": ", ".join(config.REVIEWS_GITHUB_REPOSITORY_CONFIGURATION),
+            "value": ", ".join(settings.REVIEWS_GITHUB_REPOSITORY_CONFIGURATION),
         },
         {
             "name": "REVIEWS_GITLAB_REPOSITORY_CONFIGURATION",
-            "value": ", ".join(config.REVIEWS_GITLAB_REPOSITORY_CONFIGURATION),
+            "value": ", ".join(settings.REVIEWS_GITLAB_REPOSITORY_CONFIGURATION),
         },
         {
             "name": "REVIEWS_LABEL_CONFIGURATION",
-            "value": ", ".join(config.REVIEWS_LABEL_CONFIGURATION),
+            "value": ", ".join(settings.REVIEWS_LABEL_CONFIGURATION),
         },
     ]
