+++ reviews/config/__init__.py
@@ -1,5 +1,4 @@
-from ..config.helpers import get_configuration, get_label_colour_map  # NOQA: F401
-from ..config.render import render_config_table  # NOQA: F401
+from ..config.helpers import get_configuration  # NOQA: F401
 from ..config.settings import (  # NOQA: F401
     GITHUB_DEFAULT_PAGE_SIZE,
     GITHUB_TOKEN,
